moved populateLandButtons to HomeFragment.java. further implemented populateLandButtons method.
