+++ app/src/main/java/com/farmgame/Activity/main/HomeFragment.java
@@ -4,6 +4,9 @@ import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.TableLayout;
+import android.widget.TableRow;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
@@ -17,7 +20,8 @@ import com.farmgame.databinding.FragmentHomeBinding;
 
 public class HomeFragment extends Fragment {
 
-
+    private static final int ROW_NUM = 4;
+    private static final int COL_NUM = 5;
     private FragmentHomeBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
@@ -28,10 +32,27 @@ public class HomeFragment extends Fragment {
         View root = binding.getRoot();
 
         final TextView textView = binding.textHome;
-
         return root;
     }
-
+    private void populateLandButtons() {
+        for(int row = 0; row < ROW_NUM; row ++){
+            TableLayout table = binding.tableForLandButtons;
+            TableRow tableRow = new TableRow(requireActivity());
+            tableRow.setLayoutParams(new TableLayout.LayoutParams(
+                    TableLayout.LayoutParams.MATCH_PARENT,
+                    TableLayout.LayoutParams.MATCH_PARENT,
+                    1.0f));
+            table.addView(tableRow);
+            for (int col = 0; col < COL_NUM; col ++){
+                Button button = new Button(requireActivity());
+                tableRow.addView(button);
+                button.setLayoutParams(new TableRow.LayoutParams(
+                        TableRow.LayoutParams.MATCH_PARENT,
+                        TableRow.LayoutParams.MATCH_PARENT,
+                        1.0f));
+            }
+        }
+    }
     @Override
     public void onDestroyView() {
         super.onDestroyView();
